code review and using ze types

+++ include/imp/cu_core/cu_texture2d.cuh
@@ -34,7 +34,7 @@ struct Texture2D
   {
   }
 
-  __host__ Texture2D(const void* data, size_t pitch,
+  __host__ Texture2D(const void* data, uint32_t pitch,
                      cudaChannelFormatDesc channel_desc,
                      ze::Size2u size,
                      bool _normalized_coords = false,
